+++ src/index.d.ts
@@ -1,3 +1,0 @@
-import * as fnUtil from './fnUtil.js';
-export { fnUtil };
-//# sourceMappingURL=index.d.ts.map