+++ vite.config.js
@@ -23,4 +23,12 @@ export default defineConfig({
             },
         },
     },
+    test: {
+        coverage: {
+            provider: 'v8',
+            reporter: ['text', 'json', 'html', 'lcov'],
+            include: ['src/**/*.ts'],
+            exclude: ['src/**/*.d.ts', 'src/index.ts'],
+        },
+    },
 });